refactor(runtime): ♻️  Organize & remove more occurencies of `RuntimeError`
