+++ .changeset/shaggy-bugs-unite.md
@@ -1,0 +1,5 @@
+---
+"@terminal-nerds/snippets-runtime": patch
+---
+
+♻️ Remove more occurences of `RuntimeError` to fix CJS issues
